+++ docs/changelog.rst
@@ -16,6 +16,7 @@ Fixed
 
 - ``ValueError`` during sending a request with test payload if the endpoint defines a parameter with ``type: array`` and ``in: formData``. `#661`_
 - ``KeyError`` during processing a schema with nullable parameters and ``in: body``. `#660`_
+- ``StopIteration`` during ``requestBody`` processing if it has empty "content" value. `#673`_
 
 `2.2.1`_ - 2020-07-22
 ---------------------
@@ -1261,6 +1262,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#673: https://github.com/kiwicom/schemathesis/issues/673
 .. _#665: https://github.com/kiwicom/schemathesis/issues/665
 .. _#661: https://github.com/kiwicom/schemathesis/issues/661
 .. _#660: https://github.com/kiwicom/schemathesis/issues/660